+++ client/src/App.js
@@ -12,6 +12,7 @@ import Ops from './components/Ops'
 import './App.css'
 import { loadCategories, loadSamples, loadTests } from './reducer'
 import { getCategories, getSamples, getTests } from './utils/request'
+import TestList from './containers/TestList'
 
 export default function App() {
   const dispatch = useDispatch()
@@ -26,6 +27,7 @@ export default function App() {
       <Routes>
         <Route path='/' element={<Home />} />
         <Route path='/home' element={<Home />} />
+        <Route path='/tests' element={<TestList />} />
         <Route path='/signup' element={<Signup />} />
         <Route path='/about' element={<About />} />
         <Route path='/contact' element={<Contact />} />
